improve `ArcCow`
